requirements-dev.txt
setuptools
wheel
pytest
cpufeature
sphinx
pydata_sphinx_theme
gcovr
tensorflow
torch
scipy
pyyaml
mypy